refactor: move atto parser implementation into its own module
